Add reference title mapping
This will add child `reference` elements to `Benchmark` element in
XCCDF. The value of these `reference` elements will be the title or type
of the reference and their `href` attribute will be set to the reference
URI. This way we will have a mapping from a reference URI to reference
title stored in our content. In future, we can use this mapping for
easier work with references in OpenSCAP, oscap-report, etc.

This commit also adds a small test that ensures that this reference
mapping in the built XCCDF file is the same as the contents of the
`reference_uris` key in resolved product YAML
(`build/${PRODUCT}/product.yml`).